refactor(app): split the two largest chart files into directories
  HDXMultiSeriesTimeChart.tsx  1503 -> 720 (+ 8 files)
  DBTimeChart.tsx              1022 -> 469 (+ 6 files)

Each becomes a directory with a barrel, so every import path and every
jest.mock('@/...') keeps resolving and the change is invisible to consumers.

The seams follow what the code already separated rather than cutting by length:

- searchUrl.ts takes buildSearchUrl out of a useCallback as a pure function,
  which makes its branching testable for the first time — which value column a
  series key resolves to, and whether that column's aggregation is attributable
  to individual events at all. A non-attributable aggregation must not produce a
  value filter, or drill-down returns rows that never contributed to the clicked
  point. Nine new tests cover it; that is the only new test surface here.
- useChartScales holds the axis domains and the annotation elements. Pure
  derivation from props, no state, no recharts tree.
- The tooltip, legend, recharts shape shims, layout constants, cross-chart pin
  registry and pure data helpers each move to a file named after what they are.

MemoChart stays at 720 because what is left is one recharts element whose
children must remain siblings, plus interaction state that genuinely shares a
click-suppression flag with the brush-zoom. Splitting that tree would trade a
real risk of dropping a render branch for a smaller number.

Two deliberate changes beyond pure movement, both small:

- xAxisDomain is typed as the [number, number] tuple it already returned rather
  than the wider AxisDomain, which removes two unsafe type assertions at its
  consumers.
- dismissPinned and buildSearchUrl became block-bodied and a thin delegate
  respectively. Same behaviour.

Verified by diffing every non-import line of the two originals against the new
directories: the only lines that do not appear are import fragments that were
redistributed and the two changes above.

make ci-lint, make ci-unit (5205 tests) and the dashboard E2E suite (81 passed)
all pass. E2E matters most here — it is the only check that catches a
runtime-only breakage from moving code.

First of the pieces split out of #2536. No feature code: this lands ahead of the
exemplars overlay so that PR is a feature diff rather than a feature plus a
1,500-line move.
