+++ packages/app/src/components/DBTimeChart/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Public surface of the time-chart tile. Split out of a single 1000-line module;
+ * consumers (and the test files that `jest.mock('@/components/DBTimeChart')`)
+ * import from here, so the internal file layout stays free to change.
+ */
+export { DBTimeChart } from './DBTimeChart';
+export { decodeSeriesGroupFilters, type SeriesGroupFilter } from './searchUrl';
